#11 - 캘린더 화면 리팩토링
더이상 사용되지 않는 위젯들을 정리하였다.
