+++ lib/screen/main/tab/calendar/w_personal_calendar.dart
@@ -1,29 +1,47 @@
 import 'package:flutter/material.dart';
+import 'package:playground/screen/main/tab/calendar/vo/vo_user_schedule.dart';
 import 'package:table_calendar/table_calendar.dart';
 
 class PersonalCalendar extends StatefulWidget {
   const PersonalCalendar({
     super.key,
-    required DateTime focusedDay,
-    required DateTime firstDay,
-    required DateTime lastDay,
-  }) : _focusedDay = focusedDay, _firstDay = firstDay, _lastDay = lastDay;
+    required this.userSchedule,
+    required this.focusedDay,
+    required this.firstDay,
+    required this.lastDay,
+  });
 
-  final DateTime _focusedDay;
-  final DateTime _firstDay;
-  final DateTime _lastDay;
+  final UserSchedule userSchedule;
+  final DateTime focusedDay;
+  final DateTime firstDay;
+  final DateTime lastDay;
 
   @override
   State<PersonalCalendar> createState() => _PersonalCalendarState();
 }
 
 class _PersonalCalendarState extends State<PersonalCalendar> {
+
+  late UserSchedule _userSchedule;
+  late DateTime _focusedDay;
+  late DateTime _firstDay;
+  late DateTime _lastDay;
+
+  @override
+  void initState() {
+    _userSchedule = widget.userSchedule;
+    _focusedDay = widget.focusedDay;
+    _firstDay = widget.firstDay;
+    _lastDay = widget.lastDay;
+    super.initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return TableCalendar(
-      focusedDay: widget._focusedDay,
-      firstDay: widget._firstDay,
-      lastDay: widget._lastDay,
+      focusedDay: _focusedDay,
+      firstDay: _firstDay,
+      lastDay: _lastDay,
       locale: 'ko',
     );
   }
